fixed shuffling, made indigo working dir independent
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@85 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ networking/packet.h
@@ -65,7 +65,7 @@ public:
 
 	inline const char* getData() const { return &(*data.begin()); }
 	inline char* getData(){ return &(*data.begin()); }
-	inline int getPacketSize() const { return data.size(); }
+	inline int getPacketSize() const { return (int)data.size(); }
 
 	void writeToStreamSized(MyStream& stream);//write to stream with an int indicating the size first.
 	void readFromStreamSize(MyStream& stream);
